0.0.6.0 Window scale updates
Updates:
-Changed the scale variable to suport decimals (it is now a float)
-You can now change the scale of the game in-game with O, P keys. (Note, this is a basic feature to further test maps and it is still in development)

+++ Proyecto2/Motor2D/Input.cpp
@@ -160,7 +160,7 @@ bool ModuleInput::PreUpdate()
 			break;
 
 		case SDL_MOUSEMOTION:
-			int scale = app->win->GetScale();
+			float scale = app->win->GetScale();
 			mouseMotionX = event.motion.xrel / scale;
 			mouseMotionY = event.motion.yrel / scale;
 			mouseX = event.motion.x / scale;
